code cleanup

+++ app/controllers/farriers_controller.rb
@@ -1,19 +1,25 @@
 class FarriersController < ApplicationController
+  after_action :verify_authorized
+
 
   def index
     @farriers = Farrier.all
+    authorize @farriers
   end
 
   def show
     @farrier = Farrier.find(params[:id])
+    authorize @farrier
   end
 
   def new
     @farrier = Farrier.new
+    authorize @farrier
   end
 
   def create
     @farrier = Farrier.new(farrier_params)
+    authorize @farrier
     if @farrier.save
       flash[:notice] = "New Farrier Added!"
       redirect_to @farrier
@@ -25,10 +31,12 @@ class FarriersController < ApplicationController
 
   def edit
     @farrier = Farrier.find(params[:id])
+    authorize @farrier
   end
 
   def update
    @farrier = Farrier.find(params[:id])
+   authorize @farrier
    if @farrier.update_attributes(farrier_params)
      flash[:notice] = "Farrier Updated!"
      redirect_to @farrier
@@ -39,7 +47,7 @@ class FarriersController < ApplicationController
 
   def destroy
     @farrier = Farrier.find(params[:id])
-
+    authorize @farrier
     if @farrier.destroy
       flash[:alert] = 'Farrier has been deleted.'
       redirect_to new_farrier_path
@@ -59,5 +67,7 @@ class FarriersController < ApplicationController
     )
   end
 
+  
+ 
 
 end
