+++ app/controllers/problems_controller.rb
@@ -1,20 +1,33 @@
 class ProblemsController < ApplicationController
-  before_action :authorize_user, except: [:show]
-  
+  before_action :authorize_user
+
+
+
+  def index
+    @problems = policy_scope(Problem).all
+  end
+
   def show
-    @problem = Problem.find(params[:id])
-    @horse = @problem.horse
+      @problem = Problem.find(params[:id])
+      authorize @problem
+      if current_user.id != @problem.user_id
+        respond_to do |format|
+          format.json { render json: @problems }
+          format.html
+        end
+      end
   end
 
   def new
     @problem = Problem.new
+    authorize @problem
   end
 
   def create
     @problem = Problem.new(problem_params)
-    @user = User.all
+    authorize @problem
     if @problem.save
-    CreateHorseMailer.update_user_create(current_user, @user, @horse, @problem).deliver
+    NewProblemCreatedMailer.new_problem_created(current_user, @horse, @problem).deliver
       flash[:notice] = "New problem Added!"
       redirect_to @problem
     else
@@ -25,15 +38,16 @@ class ProblemsController < ApplicationController
 
   def edit
     @problem = Problem.find(params[:id])
+    authorize @problem
   end
 
 
   def update
    @problem = Problem.find(params[:id])
+   authorize @problem
    @horse = @problem.horse
-   @user = User.all
    if @problem.update_attributes(problem_params)
-    UpdateHorseMailer.update_user(current_user, @user, @horse, @problem).deliver
+    UpdateHorseMailer.update_user(current_user, @horse, @problem).deliver
      flash[:notice] = "Problem Updated!"
      redirect_to @problem
    else
@@ -43,13 +57,19 @@ class ProblemsController < ApplicationController
 
   def destroy
     @problem = Problem.find(params[:id])
-
+    authorize @problem
     if @problem.destroy
       flash[:notice] = 'Problem has been deleted.'
       redirect_to new_problem_path
     end
   end
 
+  def problems
+    @users = @user.all.collect { |u| [u.name, u.id] }
+    respond_to do |format|
+      format.json { render json: @users } 
+    end
+  end 
 
   private
 
@@ -59,15 +79,14 @@ class ProblemsController < ApplicationController
       :time,
       :description,
       :notes,
+      :horse_id,
       :user_id,
-      :horse_id
+      :id,
       )
   end
 
-
-
   def authorize_user
-    if !user_signed_in? || !current_user.admin?
+    if !user_signed_in?
       raise ActionController::RoutingError.new("Not Found")
     end
   end
