+++ app/controllers/veterinarians_controller.rb
@@ -1,18 +1,25 @@
 class VeterinariansController < ApplicationController
+  before_action :authorize_user, except: [:show]
+
+
   def index
     @veterinarians = Veterinarian.all
+    authorize @veterinarians
   end
 
   def show
     @veterinarian = Veterinarian.find(params[:id])
+    authorize @veterinarian
   end
 
   def new
     @veterinarian = Veterinarian.new
+    authorize @veterinarian
   end
 
   def create
     @veterinarian = Veterinarian.new(veterinarian_params)
+    authorize @veterinarian
     if @veterinarian.save
       flash[:notice] = "New Veterinarian Added!"
       redirect_to @veterinarian
@@ -24,10 +31,12 @@ class VeterinariansController < ApplicationController
 
   def edit
     @veterinarian = Veterinarian.find(params[:id])
+    authorize @veterinarian
   end
 
   def update
    @veterinarian = Veterinarian.find(params[:id])
+   authorize @veterinarian
    if @veterinarian.update_attributes(veterinarian_params)
      flash[:notice] = "Veterinarian Updated!"
      redirect_to @veterinarian
@@ -38,7 +47,7 @@ class VeterinariansController < ApplicationController
 
   def destroy
     @veterinarian = Veterinarian.find(params[:id])
-
+    authorize @veterinarian
     if @veterinarian.destroy
       flash[:notice] = 'Veterinarian has been deleted.'
       redirect_to new_veterinarian_path
